lesson-14/03/pyproject.toml
[build-system]
requires = [
    "hatchling"
]
build-backend = "hatchling.build"

[project]
name = "animal"
description = "A simple animal package"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: 3.15",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Unix",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: Microsoft :: Windows :: Windows 11",
]
dependencies = [
    "docopt-ng>=0.9.0, <1"
]
dynamic = [
    "version"
]

[project.license]
text = "MIT"

[[project.authors]]
name = "Author Name"
email = "user@example.com"

[project.urls]
Homepage = "https://example.com/"
Registry = "https://example.com/"

[project.scripts]
animal = "animal.cli:main"

[project.optional-dependencies]
dev = ["pytest"]

[tool.hatch.build.targets.sdist]
include = ["src/animal"]

[tool.hatch.build.targets.wheel]
packages = ["src/animal"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.version]
path = "src/animal/__about__.py"

[tool.versioningit.vcs]
method = "git"
default-tag = "0.0.0"

[tool.pytest.ini_options]
log_cli = false
log_cli_level = "INFO"
minversion = "8.4"
pythonpath = ["src"]
testpaths = ["tests"]
